Move Output functions to SvnControl
