+++ lib/git/si.rb
@@ -3,7 +3,6 @@ require "git/si/errors"
 require "git/si/svn-control"
 require "git/si/git-control"
 require "git/si/git-ignore"
-require "git/si/output"
 require "git/si/util"
 require "git/si/actions"
 require "thor"
