+++ lib/git/si/svn-control.rb
@@ -114,6 +114,21 @@ module Git
         end.compact
       end
 
+      def self.parse_svn_status(status_string)
+        return '' unless status_string
+        status_string.split(/\r?\n/).select do |line|
+          line.strip !~ /(^X|\.git|\.swp$)/
+        end
+      end
+
+      def self.parse_external_repos(status_string)
+        status_string.split(/\r?\n/).collect do |line|
+          line.strip.match(/^\s*X\s+(\S.+)/) do |pattern|
+            pattern.to_a.compact.last
+          end
+        end.compact
+      end
+
     end
   end
 end
